+++ PowerShellForWebInspect/Public/New-WIKnownTechnologyDescriptorObject.ps1
@@ -1,10 +1,18 @@
 function New-WIKnownTechnologyDescriptorObject
 {
+    <#
+    .SYNOPSIS
+        Create a new KnownTechnologyDescriptorObject.
+    .DESCRIPTION
+        Create a new PS4WI.KnownTechnologyDescriptorObject for use in initiating a new scan.
+    .FUNCTIONALITY
+        WebInspect
+    #>
     [CmdletBinding()]
     [OutputType([System.Collections.Hashtable],[String])]
     param
     (
-        [Parameter]
+        [Parameter()]
         [validateset('none', 'webSpherePortal', 'oracleADFFaces')]
         [string]$Framework,