+++ PowerShellForWebInspect/Public/New-WIMacroParameterObject.ps1
@@ -1,5 +1,19 @@
 function New-WIMacroParameterObject
 {
+    <#
+    .SYNOPSIS
+        Create a new MacroParameterObject.
+    .DESCRIPTION
+        Create a new PS4WI.MacroParameterObject for use in initiating a new scan.
+    .PARAMETER Name
+        The name of the parameter.
+    .PARAMETER Value
+        The value of of the parameter.
+    .EXAMPLE
+        $param1 = New-WIMacroParameterObject -Name "username" -Value "user"
+    .FUNCTIONALITY
+        WebInspect
+    #>
     [CmdletBinding()]
     [OutputType([System.Collections.Hashtable],[String])]
     param