+++ PowerShellForWebInspect/Public/New-WIProxyConfigurationDescriptorObject.ps1
@@ -1,33 +1,57 @@
 function New-WIProxyConfigurationDescriptorObject
 {
+    <#
+    .SYNOPSIS
+        Create a new ProxyConfigurationDescriptorObject.
+    .DESCRIPTION
+        Create a new PS4WI.ProxyConfigurationDescriptorObject for use in initiating a new scan.
+    .PARAMETER ProxyMode
+        The mode of the proxy.
+    .PARAMETER ProxyServer
+        The proxy server hostname or IP address.
+    .PARAMETER ProxyPort
+        The proxy server port number.
+    .PARAMETER ProxyPacUrl
+        The proxy server autoconfiguration PAC file URL.
+    .PARAMETER ProxyAuthType
+        The proxy server authentication type.
+    .PARAMETER ProxyAuthUsername
+        The proxy authentication username.
+    .PARAMETER ProxyAuthPassword
+        The proxy authentication password.
+    .PARAMETER ProxyBypass
+        The bypass proxy semicolon separated list.
+    .FUNCTIONALITY
+        WebInspect
+    #>
     [CmdletBinding()]
     [OutputType([System.Collections.Hashtable],[String])]
     param
     (
-        [Parameter]
+        [Parameter(Mandatory)]
         [validateset('none', 'standard', 'socks4', 'socks5', 'pacUrl')]
         [string]$ProxyMode,
 
-        [Parameter]
+        [Parameter(Mandatory)]
         [string]$ProxyServer,
 
-        [Parameter]
+        [Parameter(Mandatory)]
         [int]$ProxyPort,
 
-        [Parameter]
+        [Parameter()]
         [string]$ProxyPacUrl,
 
-        [Parameter]
+        [Parameter()]
         [validateset('none', 'basic', 'ntlm', 'kerberos', 'digest', 'automatic', 'negotiate')]
         [string]$ProxyAuthType,
 
-        [Parameter]
+        [Parameter()]
         [string]$ProxyAuthUsername,
 
-        [Parameter]
+        [Parameter()]
         [string]$ProxyAuthPassword,
 
-        [Parameter]
+        [Parameter()]
         [string]$ProxyBypass
     )
     begin