+++ PowerShellForWebInspect/Public/New-WIScanSettingsOverrideObject.ps1
@@ -1,10 +1,94 @@
 function New-WIScanSettingsOverrideObject
 {
+    <#
+    .SYNOPSIS
+        Create a new ScanSettingsOverrideObject object.
+    .DESCRIPTION
+        Create a new PS4WI.ScanSettingsOverrideObject with any custom settings overrides for use in
+        initiating a new scan.
+    .PARAMETER ScanName
+        Any alpha-numeric value, does not need to be unique.
+        Required.
+    .PARAMETER StartUrls
+        A list of valid, fully qualified URLs, including scheme, host and port.
+        This field is only used if "StartOption" is "Url".
+    .PARAMETER UserAgent
+        A UserAgentDescriptorObject which can be used to configure a predefined user agent or use a
+        custom user agent string.
+    .PARAMETER WebServiceScan
+        A WebServiceScanDescriptorObject which can be used to configure a SOAP Web Service Scan.
+    .PARAMETER CrawlAuditMode
+        The crawl and/or audit mode to use.
+    .PARAMETER CrawlCoverageMode
+        The crawl coverage mode to use.
+    .PARAMETER KnownTechnology
+        A KnownTechnologyDescriptorObject to use.
+    .PARAMETER SharedThreads
+        A number between 1 and 75 indicating the number of shared threads that the crawler and auditor should use
+        (default is 1, meaning a single threaded scan). SharedThreads is not compatible with CrawlThreads and
+        AuditThreads. If SharedThreads is set, CrawlThreads and AuditThreads will be ignored.
+    .PARAMETER CrawlThreads
+        A number between 1 and 25 indicating the number of threads that the crawler should use (default is 1).
+        SharedThreads is not compatible with CrawlThreads and AuditThreads. If SharedThreads is set, CrawlThreads
+        and AuditThreads will be ignored.
+    .PARAMETER AuditThreads
+        A number between 1 and 50 indicating the number of threads that the auditor should use (default is 10).
+        SharedThreads is not compatible with CrawlThreads and AuditThreads. If SharedThreads is set, CrawlThreads
+        and AuditThreads will be ignored.
+    .PARAMETER StartOption
+        The start option to use.
+    .PARAMETER LoginMacroAutoGen
+        A MacroGenDescriptorObject which can be used to configure automatic login macro generation.
+    .PARAMETER LoginMacro
+        A webmacro file name.
+        This file must exist in the WebInspect scan settings folder on the WebInspect machine.
+    .PARAMETER MacroParameters
+        A MacroParametersObject which can be used to override parameters defined in TruClient macros.
+        The macros must exist in the scan settings and the parameters must exist in the macro.
+    .PARAMETER SmartCredentials
+        Override the smart credentials defined in a session-based macro. The macro must exist in the scan settings
+        and smart credentials must be enabled in the macro.
+    .PARAMETER Proxy
+        A ProxyConfigurationDescriptorObject to use which defines a proxy server.
+    .PARAMETER NetworkCredentials
+        Network authentication username and password. Use the NetworkAuthenticationMode setting to specify the
+        authentication mode.
+    .PARAMETER NetworkAuthenticationMode
+        Specifies the network authentication mode when using NetworkCredentials, e..g. "myusername","mypassword"
+    .PARAMETER AllowedHosts
+        An array of allowed "host:port" entries. Hosts found in "StartUrls" are automatically added to the allowed
+        hosts list. Use the special value "*" to add all hosts found in workflow macros, login macros and start
+        urls to allowed hosts.
+        Caution! This is a convenience feature, and it can cause the scanner to go out of scope.
+        Examples: "zero.webappsecurity.com:80", "zero.webappsecurity.com:443", "myhost:8888" or "\*"
+    .PARAMETER RestEndPoints
+        A RestEndpointDescriptorObject containing the REST end points to use.
+    .PARAMETER WebFormValues
+        The web form values to use.
+    .PARAMETER PolicyId
+        An integer representing the policy id OR the GUID representing the policy unique id.
+        Use "Get-WIPolicies" for a list of available policies.
+    .PARAMETER CheckIds
+        An array of check IDs.
+        A custom audit policy will be created from this list and used for the scan.
+    .PARAMETER DontStartScan
+        If this switch is set, the scan will be created but it will remain in the stopped state.
+    .PARAMETER ScanScope
+        The scope of the scan.
+    .PARAMETER ScopedPaths
+        Restrict the crawler to specific paths within a web site.
+    .PARAMETER ClientCertificate
+        Either the location of the client certificate in the Windows certificate store OR the raw client certificate.
+    .PARAMETER EnableTrafficMonitor
+        If this switch is set, the traffic monitor will be enabled.
+    .FUNCTIONALITY
+        WebInspect
+    #>
     [CmdletBinding()]
     [OutputType([System.Collections.Hashtable],[String])]
     param
     (
-        [Parameter(Mandatory)]
+        [Parameter()]
         [string]$ScanName,
 
         [Parameter()]
@@ -59,7 +143,6 @@ function New-WIScanSettingsOverrideObject
         [PSTypeName('PS4WI.MacroParametersObject')]
         $MacroParameters,
 
-
         [Parameter()]
         [string[]]$SmartCredentials,
 
@@ -96,7 +179,6 @@ function New-WIScanSettingsOverrideObject
         [int[]]$CheckIds,
 
         [Parameter()]
-        [validateset($True, $False)]
         [switch]$DontStartScan,
 
         [Parameter()]
@@ -104,7 +186,7 @@ function New-WIScanSettingsOverrideObject
         [string]$ScanScope,
 
         [Parameter()]
-        [string[]]$ScopePaths,
+        [string[]]$ScopedPaths,
 
         [Parameter(Mandatory = $false,
                 ValueFromPipeline = $true)]
@@ -112,8 +194,7 @@ function New-WIScanSettingsOverrideObject
         $ClientCertificate,
 
         [Parameter()]
-        [validateset($True, $False)]
-        [switch]$DisableTrafficMonitor
+        [switch]$EnableTrafficMonitor
     )
     begin
     {
@@ -168,13 +249,13 @@ function New-WIScanSettingsOverrideObject
                 }
             }
             'scanScope'                 { $body.scanScope = $ScanScope }
-            'scopedPaths'               { $body.scopedPaths = $ScopePaths }
+            'scopedPaths'               { $body.scopedPaths = $ScopedPaths }
             'clientCertificate'         { $body.clientCertificate = $ClientCertificate }
-            'disableTrafficMonitor'{
-                if ($DisableTrafficMonitor) {
-                    $body.disableTrafficMonitor  = $true
-                } else {
+            'enableTrafficMonitor'{
+                if ($EnableTrafficMonitor) {
                     $body.disableTrafficMonitor  = $false
+                } else {
+                    $body.disableTrafficMonitor  = $true
                 }
             }
         }