+++ PowerShellForWebInspect/Public/New-WIWebServiceScanDescriptorObject.ps1
@@ -1,17 +1,32 @@
 function New-WIWebServiceScanDescriptorObject
 {
+    <#
+    .SYNOPSIS
+        Create a new WebServiceScanDescriptorObject.
+    .DESCRIPTION
+        Create a new PS4WI.WebServiceScanDescriptorObject for use in initiating a new scan.
+        Allows to configure SOAP Web Service Scan.
+    .PARAMETER WsdName
+        Webservice Test Design File name.
+    .PARAMETER InheritWsdProxy
+        Use Webservice Test Design File proxy settings
+    .PARAMETER InheritWsdAuth
+        Ue Webservice Test Design File network authentication settings
+    .FUNCTIONALITY
+        WebInspect
+    #>
     [CmdletBinding()]
     [OutputType([System.Collections.Hashtable],[String])]
     param
     (
-        [Parameter]
+        [Parameter()]
         [string]$WsdName,
 
-        [Parameter]
+        [Parameter()]
         [validateset($True, $False)]
         [switch]$InheritWsdProxy,
 
-        [Parameter]
+        [Parameter()]
         [validateset($True, $False)]
         [switch]$InheritWsdAuth
     )